adding bazel target
